anolis: Revert "anolis: net/smc: Fix fallback count overflow issue"
ANBZ: #7133

This reverts commit 97703974145f46745dc275107dc83e4d5082134e.

Compared with the problem itself (statistics overflow), the damage of
breaking UAPI is more serious, so revert it.

+++ net/smc/smc_stats.h
@@ -36,7 +36,7 @@ enum {
 
 struct smc_stats_fback {
 	int	fback_code;
-	u64	count;
+	u16	count;
 };
 
 struct smc_stats_rsn {
